+++ tests/unit/test_util.py
@@ -2,6 +2,7 @@
 Contains tests for utility functions.
 """
 
+import pytest
 from tests.base import TestBase
 
 import statsite.util
@@ -29,3 +30,34 @@ class TestMerge(TestBase):
 
         # Assert that it was not changed
         assert { 'a': { 'b': False } } == a
+
+class TestResolveClassString(TestBase):
+    def test_resolve_no_module(self):
+        """
+        Tests that resolving without a module fails.
+        """
+        with pytest.raises(ValueError):
+            statsite.util.resolve_class_string("Foo")
+
+    def test_resolve_no_class(self):
+        """
+        Tests that resolving without a class fails.
+        """
+        with pytest.raises(ValueError):
+            statsite.util.resolve_class_string("foo.")
+
+    def test_resolve_nonexistent_class(self):
+        """
+        Tests that resolving a class which doesn't exist
+        fails.
+        """
+        with pytest.raises(ImportError):
+            statsite.util.resolve_class_string("os.getloginn")
+
+    def test_resolve_correct_class(self):
+        """
+        Tests that resolving an item with valid parameters
+        results in a valid result.
+        """
+        result = statsite.util.resolve_class_string("os.getlogin")
+        assert callable(result)